backend/requirements.txt
uvicorn[standard]
fastapi
pydantic
python-multipart
weasyprint
pydantic-settings
requests
cryptography
PyJWT
python-docx
python-dotenv
reportlab
python-pptx
sqlalchemy>=2.0
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography] 